requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
pymongo
pymongo[srv]
certifi
dill
PyYAML
mlflow
fastapi
uvicorn

# -e .